Move dev.js -> studio.js and add rename message on 'dev' cmd
