+++ lib/plugin.js
@@ -183,6 +183,9 @@ class ServerlessEnterprisePlugin {
         enterprise: true,
         configDependent: true,
       },
+      'dev': {
+        lifecycleEvents: ['dev'],
+      },
     };
     this.hooks = {
       'login:login': this.route('login:login').bind(this),
@@ -199,6 +202,7 @@ class ServerlessEnterprisePlugin {
         this
       ),
       'studio:studio': this.route('studio:studio').bind(this),
+      'dev:dev': this.route('dev:dev').bind(this),
     };
     this.variableResolvers = {
       param: {
@@ -431,11 +435,14 @@ class ServerlessEnterprisePlugin {
           break;
         case 'studio:studio':
           if (userNodeVersion >= 8) {
-            const studio = require('./dev');
+            const studio = require('./studio');
             await studio(this);
           } else {
             this.sls.cli.log('Node 8 or higher is required to run Serverless Studio.');
           }
+          break;
+        case 'dev:dev':
+          this.sls.cli.log('The command "dev" has been renamed to "studio".');
           break;
         case 'generate-event:generate-event':
           await generate(this);
