+++ lib/studio.js
