+++ lib/studio.test.js
